Fix up the README a bit
I changed the tense in the introduction, since Fiji *does* provide an
easy ImageJ distribution (rather than only promising one in the future).

I removed the obsolete reference to NOTES in favor of mentioning that
Pull Requests are a great mechanism for participation.

I also selfishly added myself to the authors list (tautologically: I am
an author because I committed a change adding myself as an author!).

And I simplified the Contributors URL.