Descriptor-based registration: propagate single model to multiple time-points
This commit changes the behavior of the 'Descriptor-based registration'
plugin:

When a model was extracted from a pair of images with a single
time-point, the generated model could then be applied to another pair of
images. Unfortunately, it was applied only to the first time-point of
the new pair of images, even if they had several frames.

This commit allows a single model to be applied to all the time-points
of the new image pair.

The use-case targetted here is the correction for channel misalignment
in microscope able to capture several channels at once. The (typically)
two channels are often slightly misaligned. Users then image fluorescent
multi-color beads to assess this misalignement, then acquire a
time-lapse. They then want to correct the whole time-lapse from the
single image of the beads.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>