TrackMate_: DumbTracker: a tracker made for semi-automatic linking.
It is not very clever as it is based only on closest pair linking. It
tries to emulate what the user wants when he selects a bunch of spots
and click the 'link' button:
Links are created based on the existing closest pair of spots, and
created links are used to build statistics on links, which in turn are
used to forbid links that would be too far away caompared to the first
ones. It is far from from perfect, but has the advantaeg that it is
parameter-less.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>