Build.sh: handle properties directly
MiniMaven handles -D<key>=<value> command-line arguments for
convenience, setting the system properties, allowing the user to
ignore the difference between Java and MiniMaven options.

However, this support was broken in version 2.0.0-beta-7.2.

Besides, we do not really need to rely on that convenient feature
because we can be non-lazy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>