Bump all dependency versions of the dependencies in fiji.git
We do have some subprojects in fiji.git which are versioned differently
from Fiji itself, such as JEP or Skeletonize3D.

While the ultimate goal is to go to release-couplings (i.e. to rely on
release versions in dependencies to guarantee reproducible builds), for
the moment we cannot simply go back to the release versions because the
subdirectories have changed in the meantime (editing pom.xml, for
example), and therefore Jenkins would try to re-deploy them. So let's
stay with -SNAPSHOT versions until opportune moments arise to bump to a
new release version, or in the alternative to extract the projects and
put them into their own Git repositories.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>