Stitching_Grid: add "TilesToRois" checkbox
When this option is selected, all tiles in the stitched image will be
added as Rois to the RoiManager. Each will be labeled with its
appropriate Z,C,T and imageIndex coordinates in the underlying dataset,
as well as the file name on disk that the tile was read from.