Stitching_Grid: save ROIs when writing fused image
Previously, when fusing and writing to disk, ROIs were simply discarded.

The RoiManager has a command to write Rois to disk, however. So the
generateRois method and invocation were refactored to writing the final
ROIs to disk (in a "tile_rois.zip" in the output directory).

These saved ROIs can be applied after opening the saved dataset, by
opening the RoiManager and running "More >> Open...".