Report our custom-patched javac version
As reported by Curtis Rueden in bug 656,

	java -jar jars/javac-*.jar -version

did not report a valid version. The reason is the
version.properties-template file which we did not parse into a proper
version.properties file. We cannot do that anyway because we extracted
javac from OpenJDK6 b24 as per Ubuntu's sources (and applying custom
patches to address ImgLib2's overzealous use of overly-complicated
generics constructs), and therefore the JDK_VERSION would not even be
available.

So let's do the next best thing and hard-code the appropriate version
strings.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>