Add log4j SLF4J binding to fiji-compat
Dscho and I discussed, and decided that the best place to put "blanket"
dependencies for Fiji is in fiji-compat. We were already doing this e.g.
for ImageJ2 (ij-app), so it makes sense to include others there as well.
These dependencies are artifacts intended to ship with Fiji, but upon
which no specific Fiji plugin from src-plugins depends directly. By
including them as runtime dependencies of fiji-compat, we ensure that
building Fiji from source will still include these artifacts in jars/.