bf-imagej.sh: generalize most of the script
This change renames variables and reorders a few things, to centralize
Bio-Formats-specific configuration into a single "Parameters" section at
the top of the script. The idea is that in the future, this script can
be split into a more general-purpose update site deployment tool, which
can be called by e.g. "Bio-Formats 4 daily builds" and "Bio-Formats 5
daily builds" deployment jobs without violating the DRY principle.