upload-bf: remove netcdf dependency
Unfortunately, the switch from netcdf-4.0.03 to netcdf-4.0 doesn't quite
match the assumptions of the upload-update-site.sh script. If I change
the JAR to netcdf-4.0, the script will try to remove
netcdf-4.0-[0-9]*.jar, which does not quite match netcdf-4.0.03. I could
change the generic upload script to remove netcdf-4.0[0-9]* instead, but
that strikes me as a bit dangerous in case there are artifacts foo.jar
and foo-bar.jar (which we do have with e.g. ij.jar and ij-core.jar).

So for now, let's just leave netcdf off the list. It hasn't been updated
for many years (and in fact *cannot* be updated easily). So things
should be fine for the moment.