Compile a meta .jar for the externals
Without having an artifact, the dependencies of
meta-fiji-external-plugins would be ignored by the imagej-maven-plugin
(and therefore not copied to plugins/).

Let's compile a meta .jar whose only contents are a README and the
pom.xml. This documents the external plugins' versions at the same
time, killing two stones with one bird.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>