Coloc 2: Add Kendall's Tau-b rank correlation
Kendall's Tau-b rank correlation is said to be a better measure than
Spearman's rank correlation, and the latter was only preferred for a long
time because the former needs substantially more computational resources.

This implementation is very dumb and inefficient -- O(n^2) -- and is here
only for correctness.

A subsequent commit will replace it by an efficient method.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>