Introduce the PairIterator
For many measures, it should actually not be necessary to localize the
position or know about the mask. Therefore, let's abstract out an
interface for an iterator that returns pairs of values, and nothing else.

This will also make it substantially easier to test things, as one does no
longer need to construct full-blown image data structures just to test
an algorithm.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>