Add the IntArraySorter class
This class implements an introsort of an integer array given a custom
comparator. That allows users to sort index arrays, i.e. arrays
containing indexes into other arrays, and comparing the values the
indexes point to.

This is often more efficient than having full-blown Object arrays (with
the full runtime and space cost of constructing that many Objects).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>