Update Series Labeler to work with ImageJ 1.48r and later
ImageJ 1.48r removed the public class ij.plugin.frame.Fonts without
prior notice (and with poor documentation, making it a little tedious
to follow the rationale as well as to come up with a resolution).

Unfortunately, it is simply impossible to keep this plugin working
with versions prior to 1.48r because ImageJ 1.48r is simply
backwards-incompatible with every version before it due to the change
breaking the Series Labeler.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>