Build.sh: Temporarily bump MiniMaven version to -SNAPSHOT version
The long-term plan for MiniMaven is to live in its own repository, being
versioned independently of ImageJ. Once that is the case, we can easily
bump MiniMaven to a new version and use it in Build.sh when critical bugs
were fixed, such as the one Jazz Jenkins identified today. For the moment,
though, let's just go for a snapshot version and fix accordingly once
MiniMaven lives in its own repository.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>