Avoid overriding the legacy initializer
A much better way is to let LegacyInitializer do its job and declare
a service (that we might want to use for more Fiji-specific stuff
later) that the SciJava Context will discover and initialize.

That way, Fiji gets handled properly even when started from a patched
ij.jar written out by "write-headless-jar --no-headless --full-jar"
(because we no longer have to override things before starting the
LegacyEnvironment, or for that matter, patched ij.ImageJ).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>