Added a conversion to 8-bit
As the original image is only white and black, an 8 bit datatype is
sufficient to store information about the picture. Changing the datatype
allows for the use of the Glasbey spectrum. When applied to a black and
white picture, the Glasbey spectrum will assign each shade of grey a
vastly different color, allowing for easy differentiation between
shades.