Bump TrackMate version to a -SNAPSHOT to avoid jama problem
We recently bumped the Jama version used by Fiji to 1.0.3 because
part of the dependencies already used that version. However, we
did not properly prepare for that bump because the groupId (jama)
was invalid and needed to be set as a property, too.

So this is kind of a hotfix.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>