wiki-monitor: handle self-signed https:// conncections
Some Wiki -- e.g. Micro-Manager's -- insist on using HTTPS albeit with
a self-signed certificate. The symptom is a long stack trace ending in

	SunCertPathBuilderException: unable to find valid certification
	path to requested target

To remedy this, download and run

	http://wiki.openkm.com/index.php/File:InstallCert.java

(e.g. java InstallCert micro-manager.org:443) and accept the appropriate
cert (typically, the first one is the correct one, so hitting the Return
key is enough).

This will generate a jssecacerts file, and thanks to the change
introduced by this commit, the wiki-monitor will pick it up.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>