Another casualty from the switch to externalized components
The 'ready-to-upload' script was meant to help some Fiji developers to
check whether a given component is ready to be uploaded to the Fiji
update site. It is highly dubitable that this script was ever run by
the intended audience, and besides, it won't work any longer because
the source code for the Fiji components is no longer maintained in one
big, honking source code repository, but they live each in their own.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>