Let the Windows-specifc Jenkins node connect to Jenkins directly
The URL http://jenkins.imagej.net/ gets routed through Apache, the
real Jenkins is running on port 8080 (and we added the reverse proxy
only for Albert Cardona who is stuck behind a firewall that prevents
connecting to any port 8080, a firewall he cannot pinch). But our
Windows slave can connect directly. So it should do so.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>