Exclude mapdb and udunits from the updater URLs
The URLs for use in the updater bootstrapper are determined from the
dependencies on the update site itself. However, since imagej-updater
depends on imagej-common, which in turn depends on scifio, we would end
up with dependencies that are actually not needed for the updater to run
(not even for the updater GUI). So let's exclude them from the update
URLs explicitly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>