Fix_Updater.js: avoid installing the updater right away
There is a special mode in the ImageJ updater GUI that allows to update
the updater itself. To prevent an eternal loop in such a case, we have a
cunning hack in place: the thread name is set specifically to indicate
that we're currently updating the updater and that there is no need to
update it.

Since the Fix_Updater.js script is intended to run a known-good version
of the updater, let's not enforce it to be downloaded and installed, but
let's run the updater from the remote URLs instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>