Fix_Updater.js: optionally out out of adding the Fiji update site
The Fix_Updater.js script is available as
http://update.imagej.net/bootstrap.js and is the most convenient way to
bootstrap Fiji in, say, Jenkins jobs.

It is actually too usable not to allow bootstrapping ImageJ2 without the
Fiji plugins. So let's add an option to opt out from adding the Fiji
update site:

	jrunscript -Dskip.fiji=true bootstrap.js update-force-pristine

This change was made for Mark Tsuchida to assist in setting up an update
site for Micro-Manager: http://micro-manager.org/.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>