Readd missing Fiji components
Three of these were removed as unmaintained:
763e70578e6823abe7c3601b331799ffecadc030

Because they are straightforward, they will continue to be minimally
maintained by the Fiji maintainers (i.e., hinerm and me) for now.
This means that we'll keep the projects up-to-date and building, and
accept PRs, but not pursue new development or bug-fixes ourselves.

The other four are related to the ImgLib1 project. The public consensus
is to keep ImgLib1, and downstream projects like TrakEM2, in core Fiji,
at least for now. See this thread for the relevant discussion:

  https://groups.google.com/d/topic/fiji-devel/KObyQZ0R_1Q/discussion

There are issues filed to remind that the remaining ImgLib1-based
functionality should be migrated to ImgLib2:

- https://github.com/fiji/Graph_Cut/issues/2
- https://github.com/fiji/Feature_Detection/issues/1
- https://github.com/fiji/Stitching/issues/22

However, there is also TrakEM2, as well as a paper that describes how
to write (ImgLib1-based) scripts using TrakEM2. So fully removing the
ImgLib1 library from core Fiji would break backwards compatibility.

In the future, we could create a "Fiji-Legacy" update site that contains
all the currently old/deprecated functionality, so that all backwards
compatibility breakages happen at once, and are easily resolved by
enabling that update site. However, this idea needs further
consideration and discussion first, since it is arguable whether any
advantages of such a change would outweigh the disruption to end users.