Add BigDataViewer projects to core Fiji
This is achieved by importing the pom-bigdataviewer Bill Of Materials.

Care must be taken to keep the version of pom-bigdataviewer "in sync"
with the version of the pom-fiji parent project. That is: the major
digit of pom-fiji extended by pom-bigdataviewer must be the same as the
major digit of pom-fiji extended by this project. Otherwise, they will
bring in divergent dependencies somewhere in the tree structure.

See:
http://fiji.sc/Fiji_contribution_requirements#Versioning_and_dependency_convergence