Declare slf4j implementation
As an application-level project, Fiji should declare a single slf4j
logging implementation.

As Bio-Formats uses logback, we will use logback in Fiji as well. It is
added as scope:optional to avoid it being forcibly inherited by
downstream projects.

We can remove the log4j dependency, as anything with a direct log4j
dependency will either pull it in transitively (e.g. EventService) and
simply log nothing if log4j is not found.

As we are not using log4j, we do not want the slf4j-log4j adapter layer
either, so this dependency is removed.