Bump parent pom version
Updated to pom-fiji 6.1.3.

This includes a fix to the enforcer duplicate classes, excluding
com.sun:tools classes from consideratoin (which conflict with the fiji
javac fork).

With this change we can completely remove all custom enforcer
configuration from the Fiji-level pom.

Further, this should resolve build failures on systems that do contain a
tools.jar in the expected location.