ImageJ.sh: add deprecation notice
This version of the script is out of date, and broken in a few different
ways. However, it is also used by many of the utilities in Fiji's bin/
folder, as well as by (the also-deprecated) Build.sh script.

Rather than untangle all of that, for the time being, let's leave this
old version here, but with a link to the newer version in imagej.git.