upload-bf.sh: remove commented out sections
It is not quite accurate to describe the third party dependencies
as "excluded"... rather, they are all already present on an
update site (either ImageJ, Fiji or Bio-Formats). They do not
change with new builds of Bio-Formats. In other words: it is not
the concern of this script to worry about them at all.