upload-bf.sh: switch to Bio-Formats 5.1.x branch
The development of Bio-Formats 5.0.x has essentially ceased, and it is
now missing several high-impact bug-fixes and new features, with no
plans to backport them. So let's make the Bio-Formats update site start
tracking the active Bio-Formats 5.1.x development line instead.

In order to facilitate this, the update site now ships updated versions
of the three Fiji components which use Bio-Formats directly:

* SPIM_Registration
* Stitching
* legacy-imglib1

The code for these updates lives on the ome-5.1 branch in each
respective repository, as follows:

* https://github.com/fiji/SPIM_Registration/compare/ome-5.1
* https://github.com/fiji/Stitching/compare/ome-5.1
* https://github.com/fiji/legacy-imglib1/compare/ome-5.1

Once OME 5.1.0 is released, we can finalize those branches,
pinning to the release version of the Bio-Formats libraries.