upload-update-site.sh: add update site as needed
When a new Fiji is bootstrapped, the update site actually needs to be
added before doing a pristine update. Otherwise, the script will
actually _remove_ any extra files that were present on the update site,
but not on the core ImageJ and Fiji sites.

This is exactly what happened to Bio-Formats when its workspace was
deleted. To avoid it happening again, let's be sure to explicitly
(re)add the update site being updated when rebuilding the installation.

It is not actually necessary to rebuild the installation every time,
though, since the update-force-pristine option clears out any obsolete
and/or untracked files. In other words: at least in theory (assuming the
updater is bug-free), we can continually reuse the previously downloaded
and configured Fiji installation.