Stop copying JARs into jars/ and plugins/
Back in the good old days, you would build Fiji from source and it would
mutate your source directory to be an actual Fiji installation, complete
with launcher in the root directory, and all the built jars & plugins in
those respective folders.

But these days, we use Maven to build, not the old Build.sh script nor
the old Fiji Fake build system. So you end up with stuff in the target/
folder. But thanks to the imagej-maven-plugin, you can populate an
ImageJ installation with all the Fiji artifacts by passing:

    -Dimagej.app.directory=/path/to/my/Fiji.app \
    -Ddelete.other.versions=true

And all the artifacts associated with this Fiji version will be copied.

The POM was setting that imagej.app.directory property to the root
directory of the source, to partially simulate the old behavior of
Build.sh. But it does not do a complete job, and it makes for needless
extra maintenance burden with the .gitignore rules.

So, let's discontinue this feature, since it is not really useful now.