WIP: Wrap the BDV in a JavaFX window
And listen to multi-touch events.

It mostly works, although you must (?) click on the BDV canvas
before Zoom and Rotate events start firing.

I also tried with a JFXPanel inside a JFrame, which also did not work,
as discussed in this thread:

    https://community.oracle.com/thread/3578491

See also this gist, which offers another working example of
embedding Swing in a JavaFX window and receiving these events:

    https://github.com/Inervo/BookNaviger/issues/45