Adding cividis LUT
This is a mathematically optimized LUT, specifically taking into account users with a color vision deficiency. A paper is coming out soon with more details on the development of this map: Nuñez J, Anderton C, and Renslow R. Optimizing colormaps with consideration for color vision deficiency to enable accurate interpretation of scientific data. 2018.