Migrate old compatibility code from fiji-compat
In a perfect world, Fiji would not add any additional patches or hacks
on top of ImageJ2. But in reality today, these hacks are still done on
top in a production Fiji installation. In the interest of stability,
we'll keep them in place for now. But to increase the visibility of
these hacks, they are being merged into fiji/fiji from fiji/fiji-compat,
so that anyone looking at Fiji can clearly see they are happening.

This commit migrates only the still-active code. Deprecated and
unused parts have been excised, with the exception of fiji.Main,
which is still referenced by the ImageJ Launcher (!).