Update to pom-scijava 22.1.0; add HDF5_Vibez
The major goal of these updates is to bring in new-enough dependency
versions such that all are declared correctly with matching version
properties, and therefore testable with the melting-pot.sh script.

In particular, some components had a transitive dependency on
languagesupport, but with version ${rsyntaxtextarea.version}.
However, the ryntaxtextarea version (2.6.1) diverged from the
latest languagesupport version (2.6.0), resulting in an error
about languagesupport version 2.6.1 being unavailable.

These days, every single managed component in pom-scijava has
its own version property that can be individually overridden,
to avoid this sort of entanglement.

This commit brings in the following dependency version updates:

* imagej-ops: 0.41.0 -> 0.41.1
* metakit: 5.3.0 -> 5.3.1
* FS_Align_TrakEM2: 2.0.2 -> 2.0.3
* Fiji_Developer: 2.0.6 -> 2.0.7
* KymographBuilder: 2.0.0 -> 2.0.1
* OMEVisual: 1.1.1 -> 1.1.2
* Reconstruct_Reader: 2.0.3 -> 2.0.4
* Stitching_: 3.1.5 -> 3.1.6
* TrakEM2_: 1.0i -> 1.0j
* TrakEM2_Archipelago: 2.0.2 -> 2.0.3
* blockmatching_: 2.1.2 -> 2.1.3
* register_virtual_stack_slices: 3.0.3 -> 3.0.5

And the following new dependency:

* HDF5_Vibez: 1.0.1