Travis: regenerate wiki tables automatically
This means that the imagej.net wiki will always have component tables
and sidebars corresponding to what's on fiji/fiji master. This is not
perfect, because e.g. if a new component version of net.imagej:imagej
is released, that version will not be reflected on the wiki until the
fiji/fiji Travis build runs again.

For now, this is not a huge problem from a pragmatic perspective.
But if a second distribution of ImageJ starts to coexist on imagej.net
along with Fiji, now there will be fighting if that distribution also
wants to use this mechanism to update its stats and tables.