Add standard deviation projection method
Added Standard deviation option under maximum projection methods. This is particularly useful for timeseries experiments. Standard deviation option ensures pixels with large changes in intensity are brighter and thus enhanced in the final image with the temporal colour code overlaid. This is particularly useful if there is a lot of background or if surrounding objects are brighter.