bin/melt.sh: exclude org.scijava:jep from build
It has a test failure due to floating point rounding issues (I think).
But it is a retired component, and it works how it works. No reason to
keep running up against this failing test forever, nor cut a new release
that wouldn't actually change any behavior (it would only fix the test).