populate-app: install all platforms' native libs
At the moment, the only such native library is FFMPEG.

This logic was written to be general, but unfortunately due to the
variance in naming of platform-specific native JARs, it will fall down
for naming schemes other than the one hardcoded here (JavaCpp flavor).
But for the moment, it's good enough to get the job done.