populate-app: evolve along with the JogAmp version
The new JogAmp libraries do not have -main aggregator artifacts
depending on all platforms. So Fiji now depends cleanly only on the
native libraries of its current platform. That means we don't need to
clean out extraneous native library JARs anymore. But it also means we
need to explicitly populate these native libraries for platforms other
than the current one, if we want this script to populate all platforms.
Fortunately, the logic for doing that was already present for the
JavaCPP-based FFMPEG; the only problem was that the native classifier
naming scheme is different for JavaCPP versus JogAmp. So we simply add
some case logic here based on the groupId. This also makes the script
more robust in that it will warn the user if any other native libraries
are in play that do not conform to these known cases.