populate-app: deal with the multiple antlr JARs
The newer one is groupId org.antlr with package prefix org.antlr.
The older one is groupId antlr with package prefix antlr.

So both can be on the classpath... but SciJava tooling does not
like the conflicting filenames. So we rename the older one to a
more verbose groupId.artifactId-version.jar naming scheme.

If I had more time, I would do something less hacky: sort all
dependencies by artifactId and look for any duplicates, and
process them in an automated way. But since antlr is the only one
as of this writing, we do this quick and dirty hardcoded solution.