Add shim script to detect SciJava Ops failures
If an OpEnvironment input parameter is specified in a module,
but SciJava Ops is not installed, that parameter will be treated as an
*ImageJ Ops* OpEnvironment, not get handled by any preprocessor, and
then the message "An OpEnvironment is required but none exists." will be
shown to the user. This new SciJavaOpsFallback preprocessor improves on
that situation by detecting it, and giving a helpful hint on how to
install SciJava Ops so that such parameters can work successfully.

Closes #354.