WIP: Let Fiji in Python mode support Qt
With this, Python scripts from Fiji's script editor can utilize
Qt-powered software such as ndv and napari, as long as care is
taken to queue their GUI-related function calls to the main thread
(which can be done easily using superqt's ensure_main_thread logic).

However, a change is still needed to Jaunch to allow greater
configurability when launching programs on macOS: we should not
*always* start the CoreFoundation/AppKit event loop, but instead
only do so when a new directive (GUI-LOOP?) is given. Then, Fiji's
configuration in fiji.toml can include the GUI-LOOP directive when
launching in JVM mode, but *not* include it when launching in
PYTHON mode, because the fiji.py script now takes care of starting
the loop via the app.exec() call.

Before committing to this design fully, I'd like to better
understand what might be needed to make wxPython applications also
work from Fiji's Script Editor on macOS. Does wx have its own event
loop call that needs to be used? How (if at all) do people combine
Qt and wx GUI elements in complex Python programs?

Also in need of careful thinking here is scyjava's assumption that
interactive mode on macOS cannot work. Actually it can work, as
long as: 1) it launches on a non-main thread; and then 2) the
main thread runs the event loop. Not necessarily in that order,
either. With that in mind: should scyjava allow interactive mode
as long as the current thread is not the main thread? That seems
reasonable, but still, the caller needs to know to run the event
loop on the main thread; if they don't, the Fiji GUI won't appear.
How "smart" and "helpful" should we try to be about this issue?