Let Fiji in Python mode support Qt
With this, Python scripts from Fiji's script editor can utilize
Qt-powered software such as ndv and napari, as long as care is
taken to queue their GUI-related function calls to the main thread
(which can be done easily using superqt's ensure_main_thread logic).

Note that this approach requires at least Jaunch 2.0.0, which has
smarter configurability and defaults regarding launches on macOS:
Jaunch no longer always starts the CoreFoundation/AppKit event loop,
but instead does so when issued the new RUNLOOP directive -- and by
default when launching JVM applications, but not Python applications.

Still to solve: Adjust scyjava's assumptions about whether
interactive mode on macOS can work. It can work as long as:

1) it launches on a non-main thread; and
2) the main thread runs the event loop.

With that in mind: should scyjava allow interactive mode as long as
the current thread is not the main thread? That seems reasonable,
but still, the caller needs to know to run the event loop on the
main thread; if they don't, the Fiji GUI won't appear. How "smart"
and "helpful" should we try to be about this issue?

For the moment, we just use interactive:force -- evil but effective.